prime counting

+++ python_samples/worker/__init__.py
@@ -1,1 +1,6 @@
-comm_pipe = open('/tmp/worker_control_pipe_out', 'w')
+import os
+
+if os.path.exists('/tmp/worker_control_pipe_out'):
+    comm_pipe = open('/tmp/worker_control_pipe_out', 'w')
+else:
+    comm_pipe = None
